ProductSearchExportRepositoryTest: drop testing of SQL implementation
