+++ src/Shopsys/ShopBundle/Resources/config/services_test.yml
@@ -19,8 +19,7 @@ services:
     Shopsys\FrameworkBundle\Model\Product\ProductDataFactoryInterface:
         alias: Shopsys\ShopBundle\Model\Product\ProductDataFactory
 
-    Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportRepository:
-        alias: Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportWithFilterRepository
+    Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportWithFilterRepository: ~
 
     Symfony\Component\Filesystem\Filesystem: ~
